orchestrator/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
httpx